// src/type/userManagement.d.ts
export type userList = {
    key:string
    userName:string
    name:string
    email:string
    phoneNumber : number
    userId: string
}

export type addModalProps = {
    email: string
    username: string
    name: string
    country: string
    city: string
    state: string
    number: string
    dealer_id: string
    landline: string
    password: string
    pincode: string
}