// src/type/filter.d.ts
export type leadFilterProps = {
    phone: string
    name: string
    lead_status: string
    lead_id: string
    hot_lead: string
}

export interface masterFilterProps  {
  
    name:string
  
}

export interface userFilterprops  {
      userName:string
      email:string
      phoneNumber:string
      dealer_id:string
}